move category list to definitions folder
